+++ connecteur/glaneur-local/action/GlaneurLocalFicExemple.class.php
@@ -1,0 +1,19 @@
+<?php
+
+class GlaneurLocalFicExemple extends ActionExecutor {
+
+    public function go(){
+        /** @var GlaneurLocal $glaneurLocal */
+        $glaneurLocal = $this->getMyConnecteur();
+        /*
+        $result = $connecteur->recupFileExemple($this->id_e);
+        if ($result){
+            $this->setLastMessage($result);
+        } else {
+            $this->setLastMessage("Aucun fichier trouvé");
+        }
+        */
+        return true;
+    }
+
+}